Allow cluster replacement when instance type changes
When the instance type for a cache cluster changes, ensure that a
replacement cluster is created before the existing one is destroyed.
Also, encode the instance type in the cluster ID in a way that ensures
it won't go over the 20 character limit.

Lastly, set the `az_mode` attribute appropriately based on the number of
nodes in the cache cluster.